% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_session.R
\name{aoc_set_author}
\alias{aoc_set_author}
\title{Set AOC configuration options}
\usage{
aoc_set_author(author)
}
\arguments{
\item{author}{Author name, for template}
}
\value{
sets environment variable and returns success message
}
\description{
Set AOC configuration options
}
