% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_session.R
\name{aoc_set_cookie}
\alias{aoc_set_cookie}
\alias{aoc_author_set}
\alias{aoc_cookie_set}
\title{Set AOC configuration options}
\usage{
aoc_set_cookie(cookie)

aoc_author_set(author)

aoc_cookie_set(cookie)
}
\arguments{
\item{cookie}{a character string of the session cookie as found with network tools}
}
\value{
sets environment variable and returns success message
}
\description{
Set AOC configuration options
}
