% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_session.R
\name{aoc_cookie_set}
\alias{aoc_cookie_set}
\alias{aoc_cookie}
\title{Set and retrieve AOC session cookie}
\usage{
aoc_cookie_set(cookie)

aoc_cookie()
}
\arguments{
\item{cookie}{a character string of the session cookie as found with network tools}
}
\value{
sets AOC_COOKIE environment variable and returns success message

an \code{httr} object that contains a session cookie
}
\description{
Set and retrieve AOC session cookie
}
