% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_get.R
\name{aoc_get}
\alias{aoc_get}
\title{Get puzzle for day}
\usage{
aoc_get(
  day,
  year = format(Sys.Date(), "\%Y"),
  path = here::here(),
  overwrite = TRUE
)
}
\arguments{
\item{day}{the day of the puzzle}

\item{year}{year of puzzle, defaults to current year}

\item{path}{path to copy template to, defaults to subfolders of current directory}

\item{overwrite}{overwrite existing file, default to TRUE}
}
\value{
Returns the timestamp, in case you're keen on timing your results.
}
\description{
Requires AOC_COOKIE environment variable to be set. Downloads and caches puzzle inputs onto your machine.
}
