% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_init.R
\name{aoc_init}
\alias{aoc_init}
\title{Copy Template for Day}
\usage{
aoc_init(
  day,
  year = format(Sys.Date(), "\%Y"),
  path = getwd(),
  overwrite = FALSE,
  open = TRUE
)
}
\arguments{
\item{day}{the day of the puzzle}

\item{year}{year of puzzle, defaults to current year}

\item{path}{path to copy template to, defaults to subfolders of current directory}

\item{overwrite}{overwrite existing file, default to FALSE}

\item{open}{open file for editing, if RStudio is available}
}
\value{
opens the template
}
\description{
This function copies the bundled RMarkdown template into a folder within the (current) path, creating a subfolder for the year if necessary and pre-filling variables
}
