% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_read.R
\name{aoc_read}
\alias{aoc_read}
\title{Read to tibble}
\usage{
aoc_read(day, year = format(Sys.Date(), "\%Y"), convert = TRUE)
}
\arguments{
\item{day}{the day of the puzzle}

\item{year}{year of puzzle, defaults to current year}

\item{convert}{logical: whether to run \code{readr::type_convert()} or not}
}
\description{
Reads in input lines into a tibble and optionally tries type detection
}
