% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_session.R
\name{aoc_set_cookie}
\alias{aoc_set_cookie}
\alias{aoc_set_author}
\title{Set AOC configuration options}
\usage{
aoc_set_cookie(cookie)

aoc_set_author(author)
}
\arguments{
\item{cookie}{a character string of the session cookie as found with network tools}

\item{author}{Author name, for template}
}
\value{
sets environment variable and returns success message

sets environment variable and returns success message
}
\description{
AOC author will also try to read options("usethis.full_name") for the author's name, if the environment variable is not set.
}
